+++ viewmodels/LessonViewModel.php
@@ -2,24 +2,10 @@
 
 namespace ViewModels;
 
+require_once __DIR__ . '/StudentViewModels.php';
+
 use Lib\ViewModel;
-use Models\Course;
-use Models\Lesson;
-use Models\Material;
 use Models\Enrollment;
-
-class CourseView extends Course {
-    public ?string $category_name = null;
-    public ?string $instructor_name = null;
-}
-
-class LessonView extends Lesson {
-    public ?int $material_count = null;
-}
-
-class MaterialView extends Material {
-    // Additional properties if needed
-}
 
 class LessonViewModel extends ViewModel {
     /**
